requirements.txt
python-telegram-bot==20.8
asyncpg
fastapi
uvicorn
python-dotenv
